+++ libs/SeismicOperations/src/components/concrete/omp-offload/computation-kernels/isotropic/StaggeredComputationKernel.cpp
@@ -1,11 +1,22 @@
 //
-// Created by amr on 03/01/2021.
+// Created by ingy-mounir
 //
 
-#include <operations/components/independents/concrete/computation-kernels/isotropic/StaggeredComputationKernel.hpp>
+#include "operations/components/independents/concrete/computation-kernels/isotropic/StaggeredComputationKernel.hpp"
 
-#include <operations/components/dependents/concrete/memory-handlers/WaveFieldsMemoryHandler.hpp>
-#include <operations/exceptions/NotImplementedException.h>
+#include "operations/components/dependents/concrete/memory-handlers/WaveFieldsMemoryHandler.hpp"
+#include <timer/Timer.h>
+
+#include <cmath>
+#include <cstring>
+#include <float.h>
+#include <fstream>
+#include <iostream>
+#include <limits.h>
+#include <math.h>
+#include <stdio.h>
+#include <sys/time.h>
+#include <unistd.h>
 
 using namespace std;
 using namespace operations::components;
@@ -51,7 +62,6 @@ template void StaggeredComputationKernel::Compute<false, O_8>();
 template void StaggeredComputationKernel::Compute<false, O_12>();
 
 template void StaggeredComputationKernel::Compute<false, O_16>();
-
 
 template<bool IS_FORWARD_, HALF_LENGTH HALF_LENGTH_>
 void StaggeredComputationKernel::Compute() {